CustomFieldStringTemplate regular expressions now on minicard too
